+++ inc/com/centreon/engine/configuration/state.hh
@@ -1,5 +1,5 @@
 /*
-** Copyright 2011-2017 Centreon
+** Copyright 2011-2017, 2022 Centreon
 **
 ** This file is part of Centreon Engine.
 **
@@ -297,6 +297,8 @@ class state {
   void poller_id(uint32_t value) noexcept;
   uint16_t rpc_port() const noexcept;
   void rpc_port(uint16_t value) noexcept;
+  const std::string& rpc_listen_address() const noexcept;
+  void rpc_listen_address(const std::string& listen_address) noexcept;
   bool process_performance_data() const noexcept;
   void process_performance_data(bool value);
   std::list<std::string> const& resource_file() const noexcept;
@@ -563,6 +565,7 @@ class state {
   std::string _poller_name;
   uint32_t _poller_id;
   uint16_t _rpc_port;
+  std::string _rpc_listen_address;
   bool _process_performance_data;
   std::list<std::string> _resource_file;
   unsigned long _retained_contact_host_attribute_mask;