+++ src/configuration/state.cc
@@ -173,6 +173,7 @@ std::unordered_map<std::string, state::setter_func> const state::_setters{
     {"poller_name", SETTER(std::string const&, poller_name)},
     {"poller_id", SETTER(uint32_t, poller_id)},
     {"rpc_port", SETTER(uint16_t, rpc_port)},
+    {"rpc_listen_address", SETTER(const std::string&, rpc_listen_address)},
     {"precached_object_file",
      SETTER(std::string const&, _set_precached_object_file)},
     {"process_performance_data", SETTER(bool, process_performance_data)},
@@ -362,6 +363,7 @@ static bool const default_use_setpgid(true);
 static bool const default_use_syslog(true);
 static std::string const default_use_timezone("");
 static bool const default_use_true_regexp_matching(false);
+static const std::string default_rpc_listen_address("localhost");
 
 /**
  *  Default constructor.
@@ -448,6 +450,7 @@ state::state()
       _poller_name{"unknown"},
       _poller_id{0},
       _rpc_port{0},
+      _rpc_listen_address{default_rpc_listen_address},
       _process_performance_data(default_process_performance_data),
       _retained_contact_host_attribute_mask(
           default_retained_contact_host_attribute_mask),
@@ -605,6 +608,7 @@ state& state::operator=(state const& right) {
     _poller_name = right._poller_name;
     _poller_id = right._poller_id;
     _rpc_port = right._rpc_port;
+    _rpc_listen_address = right._rpc_listen_address;
     _process_performance_data = right._process_performance_data;
     _retained_contact_host_attribute_mask =
         right._retained_contact_host_attribute_mask;
@@ -754,6 +758,7 @@ bool state::operator==(state const& right) const noexcept {
       _perfdata_timeout == right._perfdata_timeout &&
       _poller_name == right._poller_name && _poller_id == right._poller_id &&
       _rpc_port == right._rpc_port &&
+      _rpc_listen_address == right._rpc_listen_address &&
       _process_performance_data == right._process_performance_data &&
       _retained_contact_host_attribute_mask ==
           right._retained_contact_host_attribute_mask &&
@@ -2113,8 +2118,8 @@ void state::host_perfdata_file_mode(perfdata_file_mode value) {
  *
  *  @return The host_perfdata_file_processing_command value.
  */
-std::string const& state::host_perfdata_file_processing_command() const
-    noexcept {
+std::string const& state::host_perfdata_file_processing_command()
+    const noexcept {
   return _host_perfdata_file_processing_command;
 }
 
@@ -2739,12 +2744,30 @@ uint16_t state::rpc_port() const noexcept {
 }
 
 /**
- *  Set poller_id value.
+ *  Set the rpc port value.
  *
  *  @param[in] value The new poller_id value.
  */
 void state::rpc_port(uint16_t value) noexcept {
   _poller_id = value;
+}
+
+/**
+ *  Get rpc_listen_address value.
+ *
+ *  @return The grpc api listen address value.
+ */
+const std::string& state::rpc_listen_address() const noexcept {
+  return _rpc_listen_address;
+}
+
+/**
+ *  Set grpc api listen_address value.
+ *
+ *  @param[in] value The new grpc api listen address.
+ */
+void state::rpc_listen_address(const std::string& listen_address) noexcept {
+  _rpc_listen_address = listen_address;
 }
 
 /**
@@ -3143,8 +3166,8 @@ void state::service_freshness_check_interval(unsigned int value) {
  *
  *  @return The service_inter_check_delay_method value.
  */
-state::inter_check_delay state::service_inter_check_delay_method() const
-    noexcept {
+state::inter_check_delay state::service_inter_check_delay_method()
+    const noexcept {
   return _service_inter_check_delay_method;
 }
 
@@ -3162,8 +3185,8 @@ void state::service_inter_check_delay_method(inter_check_delay value) {
  *
  *  @return The service_interleave_factor_method value.
  */
-state::interleave_factor state::service_interleave_factor_method() const
-    noexcept {
+state::interleave_factor state::service_interleave_factor_method()
+    const noexcept {
   return _service_interleave_factor_method;
 }
 
@@ -3235,8 +3258,8 @@ void state::service_perfdata_file_mode(perfdata_file_mode value) {
  *
  *  @return The service_perfdata_file_processing_command value.
  */
-std::string const& state::service_perfdata_file_processing_command() const
-    noexcept {
+std::string const& state::service_perfdata_file_processing_command()
+    const noexcept {
   return _service_perfdata_file_processing_command;
 }
 
@@ -3508,8 +3531,8 @@ void state::translate_passive_host_checks(bool value) {
  *
  *  @return The users resources list.
  */
-std::unordered_map<std::string, std::string> const& state::user() const
-    noexcept {
+std::unordered_map<std::string, std::string> const& state::user()
+    const noexcept {
   return _users;
 }
 
@@ -3546,7 +3569,6 @@ void state::user(std::string const& key, std::string const& value) {
 void state::user(unsigned int key, std::string const& value) {
   _users[string::from(key)] = value;
 }
-
 
 /**
  *  Set use_aggressive_host_checking value. This function is still there just